pyproject.toml
[tool.poetry]
name = "reasoning"
version = "0.1.0"
description = "A reasoning framework for all LLMs"
authors = ["colesmcintosh <colemcintosh6@gmail.com>"]
license = "MIT License"
readme = "README.md"
packages = [{include = "reasoning"}]
homepage = "https://github.com/colesmcintosh/reasoning"
repository = "https://github.com/colesmcintosh/reasoning"
documentation = "https://github.com/colesmcintosh/reasoning"
keywords = ["llm", "ai", "reasoning", "framework", "openai", "anthropic"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[tool.poetry.dependencies]
python = ">=3.9"
requests = "^2.31.0"
typing-extensions = "^4.9.0"
pydantic = "^2.6.1"
python-dotenv = "^1.0.1"
openai = "^1.61.0"
anthropic = "^0.45.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"

[tool.poetry.group.openai]
optional = true

[tool.poetry.group.openai.dependencies]
openai = "^1.61.0"

[tool.poetry.group.anthropic]
optional = true

[tool.poetry.group.anthropic.dependencies]
anthropic = "^0.45.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=reasoning --cov-report=term-missing"
